create_template

This function builds a template directory with all the necessary files, as well as a rmarkdown script which is ready to edit. The result is an Elsevier ready document that conforms to most academic journal standards. In the directory folder provided, the user will have the tex files, as well as all relevant figure folders created in R. These should not be renamed or moved. Prior to submitting a paper, the user should zip the template_name.tex file, together with the figure files and Tex folder, and send it through to journal editors. Feel free to edit any of the settings in the template_name.tex file to conform to particular preferences (e.g. setting margins, etc.)

create_template(directory = tempdir(), DoNotBuild = FALSE, template_name,
  bib.location, ShowPDFatLaunch = FALSE)

Arguments

directory
A character string with the target folder directory. This folder will be created if it does not already exist, and you will be prompted if the folder already exists. In this case, it will create a subfolder within the existing folder. By default, it will save to a temporary directory.

template_name
The folder, as well as the rmarkdown template, will be given this name.

bib.location
This can be a user specified location for a bib file. If omitted, the default bib file will be included. Note this command will create a renamed copy in the provided directory root.

ShowPDFatLaunch
FALSE by default, this will launch the default PDF.

Value

If launch_template is TRUE, the template Rmarkdown file will be opened in a new R session. This can then be used as a starting point for an academic research paper.

Examples

DoNotBuild = TRUE # Set to false to create folders structure example in your chosen directory.
create_template(directory = tempdir(), DoNotBuild = DoNotBuild, template_name = "Project")

#> 
#> Texevier Message: ========
#>  You selected not to build the folder on your computer.
#> 
#>  * Set DoNotBuild = FALSE to build the example folder in your selected directory, or
#>  * Visit the recommended folder structure, and copy the files to your preferred location, by typing: 
#> system.file(package = 'Texevier')
#> 
#> If you set DoNotBuild to TRUE, an example folder will be built in: C:\Users\NICO~1.KAT\AppData\Local\Temp\RtmpARS3V5 under your current settings
#> 
#> Change this location by setting directory = <YOURCHOSENDIRECTORY>